Optimize the display of very thin outflows
A large portion of the rendering time of the page was due to lots (sometimes hundreds) of miniature flows being drawn on top of each other. Flows that are less than a pixel in size don't need to be rendered as an area; a curved line will do. And multiple lines don't have to be drawn on top of each other; skip drawing until the shape would be out from under the stroke of the prior line.